refactor: refactor DownloadCSVButton as component
