+++ src/components/layout/Header.js
@@ -6,11 +6,9 @@ import Typography from '@material-ui/core/Typography';
 import Toolbar from '@material-ui/core/Toolbar';
 import { withStyles } from '@material-ui/core/styles';
 import CloudIcon from '@material-ui/icons/Cloud';
-import DownloadIcon from '@material-ui/icons/CloudDownload';
 import { IconButton } from '@material-ui/core';
 import TableIcon from '@material-ui/icons/TableChart';
 import RefreshIcon from '@material-ui/icons/Refresh';
-import { CSVLink } from 'react-csv';
 import { withTranslation } from 'react-i18next';
 import { ReactComponent as Logo } from '../../resources/logo.svg';
 import { DEFAULT_MODE, TEACHER_MODES } from '../../config/settings';
@@ -18,6 +16,7 @@ import { DEFAULT_VIEW, DASHBOARD_VIEW } from '../../config/views';
 import './Header.css';
 import { addQueryParamsToUrl } from '../../utils/url';
 import { getAppInstanceResources, getUsers } from '../../actions';
+import DownloadCSVButton from '../modes/teacher/DownloadCSVButton';
 
 class Header extends Component {
   static propTypes = {
@@ -31,18 +30,6 @@ class Header extends Component {
     }).isRequired,
     mode: PropTypes.string,
     view: PropTypes.string,
-    appInstanceResources: PropTypes.arrayOf(
-      PropTypes.shape({
-        data: PropTypes.string.isRequired,
-        user: PropTypes.string.isRequired,
-      })
-    ),
-    users: PropTypes.arrayOf(
-      PropTypes.shape({
-        id: PropTypes.string.isRequired,
-        name: PropTypes.string.isRequired,
-      })
-    ),
   };
 
   static styles = theme => ({
@@ -61,33 +48,12 @@ class Header extends Component {
   static defaultProps = {
     mode: DEFAULT_MODE,
     view: DEFAULT_VIEW,
-    appInstanceResources: [],
-    users: [],
   };
 
   handleRefresh = () => {
     const { dispatchGetAppInstanceResources, dispatchGetUsers } = this.props;
     dispatchGetAppInstanceResources();
     dispatchGetUsers();
-  };
-
-  downloadCSVButton = () => {
-    const { appInstanceResources, users, t } = this.props;
-
-    if (!appInstanceResources.length || !users.length) {
-      return null;
-    }
-
-    const csvData = appInstanceResources.map(({ data, user }) => {
-      const userData = users.find(({ id }) => id === user);
-      const name = userData ? userData.name : t('Anonymous');
-      return { name, data };
-    });
-    return (
-      <CSVLink data={csvData}>
-        <DownloadIcon nativeColor="#fff" />
-      </CSVLink>
-    );
   };
 
   renderViewButtons() {
@@ -98,7 +64,7 @@ class Header extends Component {
         <IconButton onClick={this.handleRefresh} key="refresh">
           <RefreshIcon nativeColor="#fff" />
         </IconButton>,
-        this.downloadCSVButton(),
+        <DownloadCSVButton />,
       ];
 
       if (view === DEFAULT_VIEW) {
@@ -143,13 +109,11 @@ class Header extends Component {
   }
 }
 
-const mapStateToProps = ({ context, appInstanceResources, users }) => ({
-  appInstanceResources: appInstanceResources.content,
+const mapStateToProps = ({ context }) => ({
   appInstanceId: context.appInstanceId,
   spaceId: context.spaceId,
   mode: context.mode,
   view: context.view,
-  users: users.content,
 });
 
 const mapDispatchToProps = {
